+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 11.0.0
 
+* `[fsutil]` `GetPerms` renamed to `GetMode`
 * `[knf]` Validators moved to sub-package (_incompatible changes_)
 * `[knf]` Removed useless dependencies
 * `[kv]` Package removed